[NR-112911] DRY up integration tests
* !jvm.isJava17Compatible() test suites should be tun w/jdk 11
* jvm.isJava17Compatible() test suites should be tun w/jdk 17

* update samples to min supported levels
