+++ gradle.properties
@@ -1,4 +1,4 @@
-org.gradle.jvmargs=-Xms512m -Xmx2048m
+org.gradle.jvmargs=-Xms512m -Xmx4096m 
 org.gradle.caching=false
 org.gradle.daemon=false
 org.gradle.warning.mode=all
