+++ plugins/gradle/build.gradle
@@ -100,7 +100,7 @@ test {
     useJUnitPlatform()
     maxHeapSize = '4G'
     onlyIf {
-        project.hasProperty('regressionTests')  // || project.hasProperty('integrationTests') ||
+        project.hasProperty('regressionTests') || project.hasProperty('integrationTests')
     }
 }
