+++ plugins/gradle/src/test/groovy/com/newrelic/agent/android/PluginDexGuardRegressionSpec.groovy
@@ -22,16 +22,9 @@ import static org.gradle.testkit.runner.TaskOutcome.SUCCESS
  */
 @Stepwise
 @IgnoreIf({ System.getProperty('regressionTests', '') != 'dexguard' })
-class PluginDexGuardRegressionSpec extends Specification {
+class PluginDexGuardRegressionSpec extends PluginRegressionSpec {
 
-    static final rootDir = new File("../..")
-    static final projectRootDir = new File(rootDir, "agent-test-app/")
-    static final buildDir = new File(projectRootDir, "build")
-    static final debuggable = false             // set to true to break in plugin/Gradle code
-
-    // Current values (update as needed)
-    static final agentVersion = "6.10.0"        // modify as needed
-    static final dexguardHome                   // = /path/to/dexguard/artifacts"
+   static final dexguardHome                   // = /path/to/dexguard/artifacts"
 
     /* According to GuardSquare, you should place your license file dexguard-license.txt
     1) in a location defined by the Java system property 'dexguard.license',
@@ -42,55 +35,11 @@ class PluginDexGuardRegressionSpec extends Specification {
     */
     static final dexguardLicenseFile            // = "/path/to/license"
 
-    @Shared
-    Map<String, String> localEnv = [:]
-
-    @Shared
-    BuildResult buildResult
-
-    @Shared
-    def testTask = 'assembleRelease'
-
-    @Shared
-    def testVariants = ['release']
-
-    @Shared
-    def printFilter
-
-    @Shared
-    String filteredOutput
-
-    // fixtures
-    def setup() {
-        printFilter = new PrintFilter()
-    }
-
-    def setupSpec() {
-        given: "Publish the agent to local M2 repo location"
-        buildResult = GradleRunner.create()
-                .withProjectDir(rootDir)
-                .withArguments("publish")
-                .build()
-
-        and: "verify M2 repo location"
-        localEnv += System.getenv()
-        if (localEnv["M2_REPO"] == null) {
-            def m2 = new File(rootDir, "build/.m2/repository")
-            if (!(m2.exists() && m2.canRead())) {
-                throw new IOException("M2_REPO not found. Run `./gradlew publish` to stage the agent")
-            }
-            localEnv.put("M2_REPO", m2.getAbsolutePath())
-        }
-    }
-
     @Unroll
     def "verify legacy #base/#plugin/#agp/#gradle instrumentation"() {
         given: "Build the app using DexGuard"
-        def runner = GradleRunner.create()
-                .forwardStdOutput(printFilter)
-                .withDebug(debuggable)
+        def runner = provideRunner()
                 .withGradleVersion(gradle)
-                .withProjectDir(projectRootDir)
                 .withArguments("--debug",
                         "-Pnewrelic.agent.version=${agentVersion}",
                         "-Pnewrelic.agp.version=${agp}",
