+++ samples/agent-test-app/gradle.properties
@@ -1,7 +1,7 @@
-org.gradle.jvmargs=-Xms256m -Xmx1024m
+org.gradle.jvmargs=-Xmx2048m -Dfile.encoding=UTF-8
 
 # configuration cache
-org.gradle.unsafe.configuration-cache=false     // FIXME
+org.gradle.unsafe.configuration-cache=false
 org.gradle.unsafe.configuration-cache-problems=WARN
 
 # build cache
@@ -9,7 +9,7 @@ org.gradle.caching=true
 
 newrelic.agent.version=7.+
 newrelic.kotlin.version=1.5.+
-newrelic.gradle.version=7.2
+newrelic.gradle.version=7.0.2
 withProductFlavors=false
 minifyEnabled=true
 compiler=r8
